fix: remove LightingBase from core
