+++ src/PepperDash.Essentials.Core/Lighting/LightingScene.cs
@@ -1,0 +1,37 @@
+﻿
+
+using System;
+using Newtonsoft.Json;
+
+namespace PepperDash.Essentials.Core.Lighting
+{
+    public class LightingScene
+    {
+        [JsonProperty("name", NullValueHandling = NullValueHandling.Ignore)]
+        public string Name { get; set; }
+        [JsonProperty("id", NullValueHandling = NullValueHandling.Ignore)]
+        public string ID { get; set; }
+        bool _IsActive;
+        [JsonProperty("isActive", NullValueHandling = NullValueHandling.Ignore)]
+        public bool IsActive 
+        {
+            get
+            {
+                return _IsActive;
+            }
+            set
+            {
+                _IsActive = value;
+                IsActiveFeedback.FireUpdate();
+            }
+        }
+
+        [JsonIgnore]
+        public BoolFeedback IsActiveFeedback { get; set; }
+
+        public LightingScene()
+        {
+            IsActiveFeedback = new BoolFeedback(new Func<bool>(() => IsActive));
+        }
+    }
+}
